+++ src/test/converter2/issues/gh2372.ts
@@ -1,0 +1,5 @@
+/**
+ * The signature for a function acting as an event handler.
+ * @param e Param should work
+ */
+export type EventHandler = (e: Event) => void;