+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Bug Fixes
 
 -   Methods which return function types no longer have duplicated comments, #2336.
+-   Comments on function-like type aliases will now show up under the type alias, rather than nested within the type declaration, #2372.
 -   Fix crash when converting some complicated union/intersection types, #2451.
 -   Navigation triangle markers should no longer display on a separate line with some font settings, #2457.
 -   `@group` and `@category` organization is now applied later to allow inherited comments to create groups/categories, #2459.